Only clean events are saved on firebase (#193)
* EventFirebaseConnection now gives default value to missing fields when adding them on Firebase.

Functions that will be used to clean collection once all produced events are clean were added too

* Fixed tests so it may comply

Not all tests could be run in a row because of TooManyRequestsException but confident they would have passed

* Fixed tests

Added a function to log in as me (melvin). The test login doesn't have a verified email and thus cannot have a profile through usual flow. We could manually put a profile but it would be error prone and need to be redone every time we change profile. With my profile if anything changes I just need to do the user flow agian and we're good